put utility functions in lib.rs
